Added Link class and updated related files for hyperlink support
A new `Link` class has been added to the `Models.cs` file to represent a hyperlink in a block of text. The `Block` class in the same file has been updated to include a list of `Link` objects. A corresponding `Link` table has been created in the database, and the `FanBentoDatabaseModelSnapshot.cs` file has been updated to include this table in the database model. The `Details.cshtml` file now supports displaying hyperlinks in text blocks. The `20240608023012_AddBlockTextLinks.Designer.cs` file has been updated to include the `Link` class in the database context. The `ProductVersion` annotation in the `FanBentoDatabaseModelSnapshot.cs` file has been updated to `8.0.6`. Navigation properties for the `Link` class have been added in the `Block` and `ContentBody` classes.